Fix stale forwarding bits bug (#777)
When copying, we erroneously assumed that the forwarding
bits of the new copy must not have been set if the forwarding
bits metadata is on the side.  That assumption was introduced in
https://github.com/mmtk/mmtk-core/commit/d341506b70a4465dbbc05e87d27d0b7bd23844bb
as an optimisation to reduce the number of forwarding bits clearing.
However, it is unsound.  If the new copy is allocated in a previously
free chunk, it may contain stale forwarding bits from the previous GC.

Now we clear forwarding bits in post_copy regardless whether the
forwarding bits metadata is in the header or on the side.

Given that post_copy always clears forwarding bits, PrepareBlockState
now only clears forwarding bits for defrag source blocks of ImmixSpace.

Closes: https://github.com/mmtk/mmtk-core/issues/776
(cherry picked from commit 4cefff7cb16c6a54d655d548f067c5f209d0a6d7)